node: support cas for batch write
